requirements.txt
# Your requirements go here
# Created by using pip3 freeze > requirements.txt
cachetools==5.3.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gspread==5.8.0
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9